[Vulkan + Profiler] Use 0 as Vulkan Event Durations During Tree Building
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89427

This change ensures that parent/child relationships between vulkan events and their corresponding CPU events are established correctly. (Previously, if a vulkan event's duration was too long, it would not be made a child correctly).

This could be merged in with the preceding diff, but I wanted to separate it for now because I'm not sure what the most appropriate way to pass through the events and adjust the in_tree_building_ flag (the way I have it now seems a bit awkward), so keeping it separate for now makes it easier to understand/fix. Taylor if you have feedback on this let me know.
ghstack-source-id: 174117514

Differential Revision: [D40084788](https://our.internmc.facebook.com/intern/diff/D40084788/)